rename entrypoint and handler packages
