+++ internal/entrypoint/promote/entrypoint.go
@@ -1,4 +1,4 @@
-package entrypoint
+package promote
 
 import (
 	"context"
