+++ internal/entrypoint/promote/mockEntrypoint.go
@@ -1,4 +1,4 @@
-package entrypoint
+package promote
 
 import (
 	"context"
