+++ internal/entrypoint/promote/options.go
@@ -1,4 +1,4 @@
-package entrypoint
+package promote
 
 // Options represents the options for the entrypoint
 type Options struct{}
